FastOr: force using bitmap containers
This change converts containers in `lazyOR` methods to bitmap containers
when computing an intersection.

It follows a similar pattern found in Java's RoaringBitmap
`RoaringBitmap.naivelazyor` method.

In our production case when using very wide _or_s this change
improves performance significantly.

When evaluating an expression which consists of two wide unions
(one of 161 bitmaps and another one of 144) the speed improvements
was of a factor of x50.

After:
```
and:30.566886ms, or:623.676526ms, andNot:10.220867ms
```

Before:
```
and:497.120619ms, or:38.185436472s, andNot:12.385317ms
```

Sadly I cannot share the input bitmaps.