Use a sync.Pool for bitmapContainers used in ToBytes()
One of our applications maintains a lot of small compressed bitmaps and memory and occasionally serializes them  to disk.  When they are serialized, we notice a large bump in memory allocations (and accompanying cpu/gc activity) due to creating a new bitmapContainer each time.

This commit modifies roaringArray to use a sync.Pool of *bitmapContainer in writeTo() instead of creating a new one each time.  Since the bitmapContainers can now be reused, a clear() method is also added to reset its state