Add Iterate() method for iterating the bitmap without an iterator
One of our applications currently makes heavy use of Iterators.  We noticed that they are a minor but noticeable source of allocations (along with the associated cpu & gc cost) in our application due to each Iterator being heap allocated along with the internal iterators each container creates.

This commit provides an alternate Iterate() method with a callback that iterates the bitmap in place, invoking the callback for each value.  This avoids allocations entirely at the cost of some code duplication